+++ arrows/DBow2/Random.cxx
@@ -28,7 +28,7 @@
 * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
-/*	
+/*
  * File: Random.cpp
  * Project: DUtils library
  * Author: Dorian Galvez-Lopez
@@ -48,7 +48,7 @@ bool DUtils::Random::m_already_seeded = false;
 void DUtils::Random::SeedRand(){
 	Timestamp time;
 	time.setToCurrentTime();
-	srand((unsigned)time.getFloatTime()); 
+	srand((unsigned)time.getFloatTime());
 }
 
 void DUtils::Random::SeedRandOnce()
@@ -62,7 +62,7 @@ void DUtils::Random::SeedRandOnce()
 
 void DUtils::Random::SeedRand(int seed)
 {
-	srand(seed); 
+	srand(seed);
 }
 
 void DUtils::Random::SeedRandOnce(int seed)
@@ -111,14 +111,14 @@ DUtils::Random::UnrepeatedRandomizer::UnrepeatedRandomizer
 int DUtils::Random::UnrepeatedRandomizer::get()
 {
   if(empty()) createValues();
-  
+
   DUtils::Random::SeedRandOnce();
-  
-  int k = DUtils::Random::RandomInt(0, m_values.size()-1);
+
+  int k = DUtils::Random::RandomInt(0, static_cast<int>(m_values.size()-1));
   int ret = m_values[k];
   m_values[k] = m_values.back();
   m_values.pop_back();
-  
+
   return ret;
 }
 
@@ -127,7 +127,7 @@ int DUtils::Random::UnrepeatedRandomizer::get()
 void DUtils::Random::UnrepeatedRandomizer::createValues()
 {
   int n = m_max - m_min + 1;
-  
+
   m_values.resize(n);
   for(int i = 0; i < n; ++i) m_values[i] = m_min + i;
 }
@@ -141,7 +141,7 @@ void DUtils::Random::UnrepeatedRandomizer::reset()
 
 // ---------------------------------------------------------------------------
 
-DUtils::Random::UnrepeatedRandomizer& 
+DUtils::Random::UnrepeatedRandomizer&
 DUtils::Random::UnrepeatedRandomizer::operator=
   (const DUtils::Random::UnrepeatedRandomizer& rnd)
 {
